completed the ui for Word of The Day and Home
for home, did a flexbox with the first div being a description, and the next few being some random, popular words.
for word of the day, added a moving vertical bar that pauses once hovered on. The function used for it was found online and uses gsap. The ads are currently a list but will be populated from the database and advertisers can get added from the advertise page
then there is a word in the top right which was done using float, however, the scrolling on this page is broken. Definitions surround the word. The word will be picked randomly from the database daily and the requirement will be that both, the word and definition, have at least 1 submitted arabic and franco version (most popular is displayed). A see more button should be added.
The doolbar is also complete, but does not support resizing. THis should be fixed and the language and darkmode button need to be linked to their respective functionality. Doing the darkmode is currenlty complicated with SASS, so it might just be a lost cause
The language theme is just what I randomly did, but someone pointed out how close it is to some other website...
Advanced Search will need to be implemented with advanced search capabilities, and home needs to be completed with likes, dislikes